Integrated react router
