+++ package.json
@@ -42,6 +42,7 @@
         "polished": "^1.9.3",
         "prop-types": "^15.6.2",
         "react-redux": "^5.0.7",
+        "react-router-dom": "^4.3.1",
         "redux": "^4.0.0",
         "redux-logger": "^3.0.6",
         "redux-thunk": "^2.3.0",
