+++ resources/assets/js/components/App.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+
+import store from '../reducers';
+import { Home } from './movies';
+
+const App = () => (
+  <Provider store={store}>
+    <Router>
+      <Switch>
+        <Route exact path="/" component={Home} />
+      </Switch>
+    </Router>
+  </Provider>
+);
+
+export default App;
