+++ resources/assets/js/components/index.js
@@ -1,15 +1,10 @@
-import React, { Component } from 'react';
-import { Provider } from 'react-redux';
+import React from 'react';
 import { render } from 'react-dom';
-
-import store from '../reducers';
-import App from './movies';
+import App from './App';
 
 if (document.getElementById('movies-app')) {
   render(
-    <Provider store={store}>
-      <App />
-    </Provider>,
+    <App />,
     document.getElementById('movies-app')
   )
 }
